git-svn-id: https://root.cern.ch/svn/root/branches/dev/mathDev@36883 27541ba8-7e3a-0410-8455-c3a389f83636

+++ bindings/pyroot/src/Converters.cxx
@@ -513,6 +513,7 @@ Bool_t PyROOT::TCStringConverter::ToMemory( PyObject* value, void* address )
    if ( fMaxSize != UINT_MAX )
       strncpy( *(char**)address, s, fMaxSize );   // padds remainder
    else
+      // coverity[secure_coding] - can't help it, it's intentional.
       strcpy( *(char**)address, s );
 
    return kTRUE;
@@ -679,7 +680,7 @@ Bool_t PyROOT::T##name##ArrayConverter::ToMemory( PyObject* value, void* address
          PyErr_SetString( PyExc_ValueError, "buffer too large for value" );  \
          return kFALSE;                                                      \
       }                                                                      \
-      memcpy( *(type**)address, buf, 0 < buflen ? buflen : sizeof(type) );   \
+      memcpy( *(type**)address, buf, 0 < buflen ? ((size_t) buflen) : sizeof(type) );\
    } else                                                                    \
       *(type**)address = (type*)buf;                                         \
    return kTRUE;                                                             \
